will this finally work?
Some abstracts and other things that might be useful later on (pt. 2).
better base robot class to replace IterativeRobot. Also we should probably really try to oop this so we don't get a mountain of spagetty code, so we may want some abstracts or other design patterny things in here.